chore: refactor the earlier/later logic
